# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: WildTails API
  version: 0.1.0
  description: >
    WildTails MVP REST API — gamified social journaling and knowledge-management platform.
    All endpoints return the unified success/error envelope.
    Private journals default to PRIVATE and are never exposed without explicit permission checks.
    AI output is always a draft; it is never published automatically.
servers:
  - url: http://localhost:3000
    description: Local development

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - RATE_LIMITED
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Human-readable message safe to display. Must not contain stack traces.
        details:
          description: Structured validation details. Never contains stack traces.
        requestId:
          type: string
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          description: Payload — shape varies per endpoint (see individual responses).
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      required: [cursor, hasMore]
      properties:
        cursor:
          type: ['string', 'null']
          description: Opaque cursor for the next page. Null when no further pages exist.
        hasMore:
          type: boolean
        total:
          type: integer
          minimum: 0
          description: Total item count — only present when explicitly requested.

    # ------ Pagination query params (inline where used) ---------------------
    # Defined here for documentation reference; used as query parameters inline.
...
    Visibility:
      type: string
      enum: [PRIVATE, PUBLIC, SELECTED_USERS, PLANET_MEMBERS]
...
    CreateJournalRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        body:
          type: string
          maxLength: 50000
          description: May be omitted for a title-only stub.
        planetId:
          type: string
          format: uuid
        visibility:
          $ref: '#/components/schemas/Visibility'
          default: PRIVATE
        tagIds:
          type: array
          maxItems: 20
          items:
            type: string
            format: uuid
        folderId:
          type: string
          format: uuid
        goalId:
          type: string
          format: uuid
...
    JournalTag:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    JournalVersion:
      type: object
      required: [id, body, createdAt]
      properties:
        id:
          type: string
          format: uuid
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    JournalResponse:
      type: object
      required: [id, title, ownerId, visibility, tags, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
          description: Omitted when caller does not have read permission.
        ownerId:
          type: string
          format: uuid
        planetId:
          type: string
          format: uuid
        visibility:
          $ref: '#/components/schemas/Visibility'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/JournalTag'
        folder:
          type: object
          required: [id, name]
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        goal:
          type: object
          required: [id, title]
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
        aiSummary:
          type: string
          description: AI-generated summary draft. Not published automatically.
        versions:
          type: array
          items:
            $ref: '#/components/schemas/JournalVersion'
          description: Previous body versions, newest first.
        sharePermissions:
          type: array
          items:
            type: object
            required: [userId, expiresAt]
            properties:
              userId:
                type: string
                format: uuid
              expiresAt:
                type: ['string', 'null']
                format: date-time
          description: Active share permissions for SELECTED_USERS journals.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ------ Knowledge — Folder ----------------------------------------------
...
    CreateSharePermissionRequest:
      type: object
      required: [journalId, userId]
      properties:
        journalId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          description: The user being granted read access.
        expiresAt:
          type: string
          format: date-time
          description: Omit for a non-expiring grant.
    SharePermissionResponse:
      type: object
      required: [id, journalId, userId, grantedBy, expiresAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        journalId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        grantedBy:
          type: string
          format: uuid
          description: Must be the journal owner.
        expiresAt:
          type: ['string', 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        revokedAt:
          type: ['string', 'null']
          format: date-time

    # ------ Social — Post ---------------------------------------------------
...
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QueryCursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque cursor from previous page. Omit for first page.
    QueryLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # -------------------------------------------------------------------------
  # Reusable responses
  # -------------------------------------------------------------------------
  responses:
    400:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    401:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    403:
      description: Authenticated but not authorised
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    404:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    409:
      description: Conflict (duplicate resource)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'

# ===========================================================================
# Tags
# ===========================================================================
...
paths:
...
  /journals:
    post:
      operationId: createJournal
      summary: Create a new journal entry
      tags: [Knowledge]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalRequest'
      responses:
        '201':
          description: Journal created (defaults to PRIVATE visibility)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/JournalResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'

    get:
      operationId: listJournals
      summary: List the authenticated user's own journals (paginated)
      tags: [Knowledge]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryCursor'
        - $ref: '#/components/parameters/QueryLimit'
        - name: folderId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by folder.
        - name: tagId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by tag.
        - name: goalId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by linked goal.
      responses:
        '200':
          description: Paginated journal list (own journals only)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/JournalResponse'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/401'
...
  /journals/{id}/share:
    post:
      operationId: createSharePermission
      summary: Grant read access to a specific user for a journal (SELECTED_USERS flow)
      tags: [Knowledge]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSharePermissionRequest'
      responses:
        '201':
          description: Share permission granted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SharePermissionResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '409':
          $ref: '#/components/responses/409'
  /journals/{id}/shares:
    get:
      operationId: listSharePermissions
      summary: List active share permissions for a journal (owner only)
      tags: [Knowledge]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Active share permissions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SharePermissionResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  /journals/{id}/share/{permissionId}:
    delete:
      operationId: revokeSharePermission
      summary: Revoke a share permission (owner only)
      tags: [Knowledge]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
        - name: permissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Permission revoked
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [revoked]
                        properties:
                          revoked:
                            type: boolean
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'

  # =========================================================================
  # Folders
  # =========================================================================
...